move tests to the right module
