+++ tests/test_fits_image.py
@@ -18,82 +18,6 @@ log = logging.getLogger("Aegean")
 log.setLevel(logging.INFO)
 
 
-def test_get_pixinfo():
-    """Test that we can get info from various header styles"""
-    header = fits.getheader('tests/test_files/1904-66_SIN.fits')
-
-    area, scale = AegeanTools.wcs_helpers.get_pixinfo(header)
-    if not area > 0:
-        raise AssertionError()
-    if not len(scale) == 2:
-        raise AssertionError()
-
-    header['CD1_1'] = header['CDELT1']
-    del header['CDELT1']
-    header['CD2_2'] = header['CDELT2']
-    del header['CDELT2']
-    area, scale = AegeanTools.wcs_helpers.get_pixinfo(header)
-    if not area > 0:
-        raise AssertionError()
-    if not len(scale) == 2:
-        raise AssertionError()
-
-    header['CD1_2'] = 0
-    header['CD2_1'] = 0
-    area, scale = AegeanTools.wcs_helpers.get_pixinfo(header)
-    if not area > 0:
-        raise AssertionError()
-    if not len(scale) == 2:
-        raise AssertionError()
-
-    header['CD1_2'] = header['CD1_1']
-    header['CD2_1'] = header['CD2_2']
-    area, scale = AegeanTools.wcs_helpers.get_pixinfo(header)
-    if not area == 0:
-        raise AssertionError()
-    if not len(scale) == 2:
-        raise AssertionError()
-
-    for f in ['CD1_1', 'CD1_2', 'CD2_2', 'CD2_1']:
-        del header[f]
-    area, scale = AegeanTools.wcs_helpers.get_pixinfo(header)
-    if not area == 0:
-        raise AssertionError()
-    if not scale == (0, 0):
-        raise AssertionError()
-
-
-def test_get_beam():
-    """Test that we can recover the beam from the fits header"""
-    header = fits.getheader('tests/test_files/1904-66_SIN.fits')
-    beam = AegeanTools.wcs_helpers.get_beam(header)
-    print(beam)
-    if beam is None:
-        raise AssertionError()
-    if beam.pa != header['BPA']:
-        raise AssertionError()
-
-    del header['BMAJ'], header['BMIN'], header['BPA']
-    beam = AegeanTools.wcs_helpers.get_beam(header)
-    if beam is not None:
-        raise AssertionError()
-
-
-def test_fix_aips_header():
-    """TEst that we can fix an aips generated fits header"""
-    header = fits.getheader('tests/test_files/1904-66_SIN.fits')
-    # test when this function is not needed
-    _ = AegeanTools.wcs_helpers.fix_aips_header(header)
-
-    # test when beam params are not present, but there is no aips history
-    del header['BMAJ'], header['BMIN'], header['BPA']
-    _ = AegeanTools.wcs_helpers.fix_aips_header(header)
-
-    # test with some aips history
-    header['HISTORY'] = 'AIPS   CLEAN BMAJ=  1.2500E-02 BMIN=  1.2500E-02 BPA=   0.00'
-    _ = AegeanTools.wcs_helpers.fix_aips_header(header)
-
-
 def test_init():
     """Test that FitsImage __init__ works """
     filename = 'tests/test_files/1904-66_SIN.fits'
